Implement BgChecker
BgChecker is a new binary used to optimize warning checking.

There are multiple client process which run specified commands a
specified frequencies and send results to the server.

The server accumulates the results and can return them immediately on
request. This makes clicking at warning button on the status bar much
more responsive.